requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docx2pdf==0.1.8
drf-spectacular==0.28.0
filelock==3.17.0
httplib2==0.22.0
identify==2.6.6
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jpype1==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
nodeenv==1.9.1
packaging==24.2
phonenumbers==8.13.54
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.1
pyreportjasper==2.1.4
pytest==8.3.4
pytest-django==4.9.0
python-docx==1.1.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
